requirements.txt
googlemaps==4.10.0
nltk==3.9.1
pandas==2.2.3
mysql-connector-python==9.1.0
numpy==2.1.3
scikit-learn==1.5.2
spacy==3.8.2